ms win routines, launchprogram.cpp (not finished)
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@2219 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/src/libafanasy/netif.cpp
@@ -92,7 +92,7 @@ void NetIF::readwrite( Msg * msg)
    rw_String(  name,    msg);
    rw_data(    (char*)macaddr, msg, MacAddrLen);
 
-   int8_t addrs = addresses.size();
+   int8_t addrs = int8_t(addresses.size());
    rw_int8_t( addrs, msg);
    for( int i = 0; i < addrs; i++)
       if( msg->isWriting()) addresses[i].write( msg);
